Cleanup fluid property materials, which are used to output fluid properties using the various formulations
