--- modules/fluid_properties/doc/content/source/materials/FluidPropertiesMaterial.md
@@ -1,12 +1,0 @@
-# FluidPropertiesMaterial
-
-`FluidPropertiesMaterial` is a base class for materials which define fluid properties
-based on a defined variable set, like `(pressure, temperature)` or `(specific volume, specific energy)`.
-
-Existing derived classes:
-
-- [FluidPropertiesMaterialPT.md]
-
-- [FluidPropertiesMaterialPH.md]
-
-- [FluidPropertiesMaterialVE.md]
