+++ modules/fluid_properties/test/tests/ideal_gas/test.i
@@ -119,7 +119,7 @@
 
 [Materials]
   [./fp_mat]
-    type = FluidPropertiesMaterial
+    type = FluidPropertiesMaterialVE
     e = e
     v = v
     fp = ideal_gas
