+++ modules/fluid_properties/test/tests/stiffened_gas/test.i
@@ -125,7 +125,7 @@
 
 [Materials]
   [./fp_mat]
-    type = FluidPropertiesMaterial
+    type = FluidPropertiesMaterialVE
     e = e
     v = v
     fp = sg
